requirements.txt
streamlit
python-dotenv
absl-py
langchain
langchain_community
langchain_google_genai
google-generativeai
faiss-cpu
torch
transformers